Add files via upload
The Advanced Statistics Project addresses three problems in various industries. It involves data analysis and hypothesis testing to provide insights on stone suitability for printing, the effectiveness of a fitness program, and the impact of dentists and implant methods on metal implant hardness. These findings will support data-driven decision-making and operational improvements for the respective companies.